+++ config/config.exs
@@ -30,6 +30,11 @@ config :lightning, LightningWeb.Endpoint,
   pubsub_server: Lightning.PubSub,
   live_view: [signing_salt: "EfrmuOUr"]
 
+config :lightning, Lightning.DistributedRateLimiter,
+  start: false,
+  capacity: 10,
+  refill_per_second: 2
+
 config :lightning, Lightning.Extensions,
   rate_limiter: Lightning.Extensions.RateLimiter,
   usage_limiter: Lightning.Extensions.UsageLimiter,