+++ lib/lightning/extensions/rate_limiting.ex
@@ -9,15 +9,11 @@ defmodule Lightning.Extensions.RateLimiting do
 
   defmodule Context do
     @moduledoc """
-    Which user is making the request for a certain project.
+    Context for the object (bucket) under rate limiting.
     """
+    @type t :: %Context{project_id: Ecto.UUID.t()}
 
-    @type t :: %Context{
-            project_id: Ecto.UUID.t(),
-            user_id: Ecto.UUID.t() | nil
-          }
-
-    defstruct [:project_id, :user_id]
+    defstruct [:project_id]
   end
 
   @callback limit_request(